refactoring + restructuring test directory + cleaning up code/logs
