+++ package.json
@@ -25,8 +25,8 @@
         "deploy:semaphore": "hardhat deploy:semaphore",
         "deploy:semaphore-voting": "hardhat deploy:semaphore-voting",
         "deploy:semaphore-whistleblowing": "hardhat deploy:semaphore-whistleblowing",
-        "test": "hardhat test",
-        "test:cross-chain": "hardhat test test/CrossChain.ts",
+        "test": "hardhat test ./test/unit_test/*",
+        "test:cross-chain": "hardhat test ./test/functional_test/CrossChain.ts",
         "test:report-gas": "REPORT_GAS=true hardhat test",
         "test:coverage": "hardhat coverage",
         "typechain": "hardhat typechain",
