+++ tsconfig.json
@@ -10,6 +10,6 @@
         "outDir": "dist",
         "typeRoots": ["node_modules/@types", "types"]
     },
-    "include": ["scripts/**/*", "tasks/**/*", "test/**/*", "build/typechain/**/*", "types/**/*"],
+    "include": ["scripts/**/*", "tasks/**/*", "test/**/*", "test/**/**/*", "build/typechain/**/*", "types/**/*"],
     "files": ["hardhat.config.ts"]
 }
